requirements_cpu.txt
# F1 Digital Twin - CPU Dependencies
# Pin ALL versions for reproducibility

# Core ML - CPU only
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu

# Numerical
numpy==1.26.4
scipy==1.12.0

# RL Environment
gymnasium==0.29.1

# Configuration
pyyaml==6.0.1
jsonschema==4.21.1

# Logging
tensorboard==2.16.2

# Testing
pytest==8.0.2
pytest-cov==4.1.0

# Type checking
mypy==1.8.0

# Code quality
ruff==0.2.2

# Utilities
tqdm==4.66.2
